+++ src/blocks/MapBlock.php
@@ -94,6 +94,16 @@ final class MapBlock extends BaseBootstrap3Block
                     ],
                 ],
             ],
+            'cfgs' => [
+                ['var' => 'snazzymapsUrl', 'label' => Module::t('block_map_snazzymapsUrl_label'), 'type' => self::TYPE_TEXT]
+            ]
+        ];
+    }
+
+    public function getFieldHelp()
+    {
+        return [
+            'snazzymapsUrl' => Module::t('block_map_snazzymapsUrl_help'),
         ];
     }
 
@@ -114,6 +124,6 @@ final class MapBlock extends BaseBootstrap3Block
      */
     public function admin()
     {
-        return '{% if vars.address is not empty %}<div class="iframe-container"><iframe src="http://maps.google.com/maps?f=q&source=s_q&hl=de&geocode=&q={{ extras.address }}&z={{ extras.zoom }}&t={{ extras.maptype }}&output=embed"></iframe></div>{% else %}<span class="block__empty-text">' . Module::t('block_map_no_content') . '</span>{% endif %}';
+        return '{% if vars.address is not empty %}<div class="iframe-container"><iframe src="{% if cfgs.snazzymapsUrl %}{{ cfgs.snazzymapsUrl }}{% else %}http://maps.google.com/maps?f=q&source=s_q&hl=de&geocode=&q={{ extras.address }}&z={{ extras.zoom }}&t={{ extras.maptype }}&output=embed{% endif %}"></iframe></div>{% else %}<span class="block__empty-text">' . Module::t('block_map_no_content') . '</span>{% endif %}';
     }
 }