+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. This projec
 
 ## 1.0.2 (in progress)
 
++ [#5](https://github.com/luyadev/luya-bootstrap3/issues/5) Added Snazzy Maps iFrame URL field to MapBlock and updated translations.
 + [#4](https://github.com/luyadev/luya-bootstrap3/pull/4) Translated FormBlock from German to English
 
 ## 1.0.1 (17. January 2018)